organized React components and containers
