+++ src/web/client/index.js
@@ -1,6 +1,6 @@
  import React from 'react';
  import ReactDOM from 'react-dom/client';
- import App from './App.jsx';
+ import App from './containers/App.jsx';
  import { BrowserRouter } from 'react-router-dom';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
